Add files via upload
To-Do Lists are actually a list of your daily tasks that you keep on your person to remind yourself of the tasks you have to accomplish that day. In this project, we are going to make a GUI based python to-do list with options to add and delete items in your list.